Likely improved performance of the defaultCompareFunction.
This adds one character to the compiled files, but it should improve the performance when the default compare function is used. Before, it checked a < b, and if false, it would check a > b. Those are two inequality checks. Now it checks a < b, and if false, it checks a == b. One inequality check and one equality check.